requirements.txt
imageio>=2.5.0
numpy
tensorflow_gpu==1.14.0
Pillow>=6.1.0

